[MOV] pos_restaurant : slip file per models
