+++ addons/pos_restaurant/models/__init__.py
@@ -1,4 +1,6 @@
 # -*- coding: utf-8 -*-
 # Part of Odoo. See LICENSE file for full copyright and licensing details.
 
+import pos_config
+import pos_order
 import pos_restaurant
