+++ addons/pos_restaurant/views/pos_config_views.xml
@@ -1,0 +1,21 @@
+<?xml version="1.0"?>
+<odoo>
+
+    <record id="view_pos_config_form" model="ir.ui.view">
+        <field name="name">pos.config.form.view.inherit</field>
+        <field name="model">pos.config</field>
+        <field name="inherit_id" ref="point_of_sale.view_pos_config_form"></field>
+        <field name="arch" type="xml">
+            <sheet position='inside'>
+                <group string="Bar &amp; Restaurant" >
+                    <field name="iface_splitbill" />
+                    <field name="iface_printbill" />
+                    <field name="iface_orderline_notes" />
+                    <field name="floor_ids" />
+                    <field name="printer_ids" />
+                </group>
+            </sheet>
+        </field>
+    </record>
+
+</odoo>
